Fix core library import path
The gedit plugin working directory isn't documented, and as such can
change from version to version. This is the root cause of #9.

This changes the core library path to be computed based on the location
of shared.py, instead of assuming the plugin working directory. This
should finally allow the plugin to use the bundled core library instead
of having to install the library separately.